+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased] - 2019-12-12
 - Add "Show Line Detail" menu command to surface action previously only discoverable via double-click.
+- Add "dateFormat" command to specify the date format used by "today" and "requireToday"
 
 ## [1.1.0] - 2019-10-19 
 - Show current script file name in window title